+++ app/client/nodeClient.ts
@@ -1,10 +1,6 @@
 import {FetchBuilder, HTTPMethod} from "~/client/base";
 import {IdRes} from "~/client/common";
 
-export interface PostRoadmapBody {
-  title: string
-}
-
 export interface PostNodeBody {
   title: string,
   description: string | undefined,
@@ -13,15 +9,7 @@ export interface PostNodeBody {
   nextNodeIds: string[],
 }
 
-export namespace roadmap {
-
-  export async function post(postRoadmapBody: PostRoadmapBody) {
-    return new FetchBuilder("/roadmap")
-      .method(HTTPMethod.POST)
-      .body(postRoadmapBody)
-      .fetch<IdRes>()
-  }
-
+export namespace NodeClient {
   export async function postNode(roadmapId: string, postNodeBody: PostNodeBody) {
     return new FetchBuilder(`/roadmap/${roadmapId}/node`)
       .method(HTTPMethod.POST)
