+++ src/main/kotlin/io/github/binaryyouchien/ensokukaido/roadmap/RoadmapRes.kt
@@ -1,0 +1,10 @@
+package io.github.binaryyouchien.ensokukaido.roadmap
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RoadmapRes(
+  val id: String,
+  val title: String,
+  val firstNodeId: String?,
+)
